+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * * A new (temporary) plugin setting `useCSS` can be set to `all` to load the old CSS for legacy browsers, or `modern` to load a polyfill.
 * Added a `fallbackImage` plugin setting (defaults to null) that will be used if no image is available for an entry.
 * Added `titleObjectTemplate` section config setting (defaults to `{title}`)
+* Added `table` layout
+* Added `size` section config setting.
 
 
 ### Enhanced customization